refactor: ticket package 분리
