+++ backend/src/main/java/com/festago/application/AdminService.java
@@ -4,14 +4,14 @@ import com.festago.domain.Festival;
 import com.festago.domain.FestivalRepository;
 import com.festago.domain.Stage;
 import com.festago.domain.StageRepository;
-import com.festago.domain.Ticket;
-import com.festago.domain.TicketAmount;
-import com.festago.domain.TicketEntryTime;
-import com.festago.domain.TicketRepository;
 import com.festago.dto.AdminFestivalResponse;
 import com.festago.dto.AdminResponse;
 import com.festago.dto.AdminStageResponse;
 import com.festago.dto.AdminTicketResponse;
+import com.festago.ticket.domain.Ticket;
+import com.festago.ticket.domain.TicketAmount;
+import com.festago.ticket.domain.TicketEntryTime;
+import com.festago.ticket.repository.TicketRepository;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
