+++ backend/src/main/java/com/festago/application/TicketingService.java
@@ -4,15 +4,15 @@ import com.festago.domain.Member;
 import com.festago.domain.MemberRepository;
 import com.festago.domain.MemberTicket;
 import com.festago.domain.MemberTicketRepository;
-import com.festago.domain.Ticket;
-import com.festago.domain.TicketAmount;
-import com.festago.domain.TicketAmountRepository;
-import com.festago.domain.TicketRepository;
 import com.festago.dto.TicketingRequest;
 import com.festago.dto.TicketingResponse;
 import com.festago.exception.BadRequestException;
 import com.festago.exception.ErrorCode;
 import com.festago.exception.NotFoundException;
+import com.festago.ticket.domain.Ticket;
+import com.festago.ticket.domain.TicketAmount;
+import com.festago.ticket.repository.TicketAmountRepository;
+import com.festago.ticket.repository.TicketRepository;
 import java.time.Clock;
 import java.time.LocalDateTime;
 import org.springframework.stereotype.Service;
