+++ backend/src/main/java/com/festago/presentation/AdminController.java
@@ -3,7 +3,6 @@ package com.festago.presentation;
 import com.festago.application.AdminService;
 import com.festago.application.FestivalService;
 import com.festago.application.StageService;
-import com.festago.application.TicketService;
 import com.festago.auth.annotation.Admin;
 import com.festago.auth.application.AdminAuthService;
 import com.festago.auth.dto.AdminLoginRequest;
@@ -15,12 +14,13 @@ import com.festago.dto.FestivalCreateRequest;
 import com.festago.dto.FestivalResponse;
 import com.festago.dto.StageCreateRequest;
 import com.festago.dto.StageResponse;
-import com.festago.dto.TicketCreateRequest;
-import com.festago.dto.TicketCreateResponse;
 import com.festago.exception.BadRequestException;
 import com.festago.exception.ErrorCode;
 import com.festago.exception.InternalServerException;
 import com.festago.exception.UnauthorizedException;
+import com.festago.ticket.application.TicketService;
+import com.festago.ticket.dto.TicketCreateRequest;
+import com.festago.ticket.dto.TicketCreateResponse;
 import io.swagger.v3.oas.annotations.Hidden;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
