+++ backend/src/main/java/com/festago/presentation/StageController.java
@@ -1,7 +1,7 @@
 package com.festago.presentation;
 
-import com.festago.application.TicketService;
-import com.festago.dto.StageTicketsResponse;
+import com.festago.ticket.application.TicketService;
+import com.festago.ticket.dto.StageTicketsResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
