+++ backend/src/main/java/com/festago/ticket/domain/TicketAmount.java
@@ -1,5 +1,6 @@
-package com.festago.domain;
+package com.festago.ticket.domain;
 
+import com.festago.domain.BaseTimeEntity;
 import com.festago.exception.BadRequestException;
 import com.festago.exception.ErrorCode;
 import jakarta.persistence.Entity;
