+++ backend/src/main/java/com/festago/ticket/domain/TicketType.java
@@ -1,4 +1,4 @@
-package com.festago.domain;
+package com.festago.ticket.domain;
 
 public enum TicketType {
     STUDENT,
