+++ backend/src/test/java/com/festago/application/integration/FestivalServiceIntegrationTest.java
@@ -7,9 +7,6 @@ import com.festago.domain.Festival;
 import com.festago.domain.FestivalRepository;
 import com.festago.domain.Stage;
 import com.festago.domain.StageRepository;
-import com.festago.domain.Ticket;
-import com.festago.domain.TicketRepository;
-import com.festago.domain.TicketType;
 import com.festago.dto.FestivalCreateRequest;
 import com.festago.dto.FestivalDetailResponse;
 import com.festago.dto.FestivalDetailStageResponse;
@@ -18,6 +15,9 @@ import com.festago.dto.FestivalResponse;
 import com.festago.support.FestivalFixture;
 import com.festago.support.StageFixture;
 import com.festago.support.TicketFixture;
+import com.festago.ticket.domain.Ticket;
+import com.festago.ticket.domain.TicketType;
+import com.festago.ticket.repository.TicketRepository;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
