+++ backend/src/test/java/com/festago/application/integration/TicketServiceIntegrationTest.java
@@ -4,20 +4,20 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.doReturn;
 
-import com.festago.application.TicketService;
 import com.festago.domain.Festival;
 import com.festago.domain.FestivalRepository;
 import com.festago.domain.Stage;
 import com.festago.domain.StageRepository;
-import com.festago.domain.TicketAmount;
-import com.festago.domain.TicketAmountRepository;
-import com.festago.domain.TicketRepository;
-import com.festago.domain.TicketType;
-import com.festago.dto.TicketCreateRequest;
-import com.festago.dto.TicketCreateResponse;
 import com.festago.exception.NotFoundException;
 import com.festago.support.FestivalFixture;
 import com.festago.support.StageFixture;
+import com.festago.ticket.application.TicketService;
+import com.festago.ticket.domain.TicketAmount;
+import com.festago.ticket.domain.TicketType;
+import com.festago.ticket.dto.TicketCreateRequest;
+import com.festago.ticket.dto.TicketCreateResponse;
+import com.festago.ticket.repository.TicketAmountRepository;
+import com.festago.ticket.repository.TicketRepository;
 import java.time.Clock;
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
