+++ backend/src/test/java/com/festago/domain/StageRepositoryTest.java
@@ -6,6 +6,9 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import com.festago.support.FestivalFixture;
 import com.festago.support.StageFixture;
 import com.festago.support.TicketFixture;
+import com.festago.ticket.domain.Ticket;
+import com.festago.ticket.domain.TicketType;
+import com.festago.ticket.repository.TicketRepository;
 import jakarta.persistence.EntityManager;
 import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
