+++ backend/src/test/java/com/festago/domain/TicketRepositoryTest.java
@@ -5,6 +5,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.festago.support.FestivalFixture;
 import com.festago.support.StageFixture;
 import com.festago.support.TicketFixture;
+import com.festago.ticket.domain.Ticket;
+import com.festago.ticket.domain.TicketType;
+import com.festago.ticket.repository.TicketRepository;
 import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
