+++ backend/src/test/java/com/festago/domain/TicketTest.java
@@ -9,6 +9,7 @@ import com.festago.support.FestivalFixture;
 import com.festago.support.MemberFixture;
 import com.festago.support.StageFixture;
 import com.festago.support.TicketFixture;
+import com.festago.ticket.domain.Ticket;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
